Add --create-signature-file option
The default behavior does not change: the file with a special name "deployed at `date` by #{committer_name}" is created in the root of the deployed repo.
If set to false, then this "signature" file is not created. In this case `--allow-empty-commits` flag becomes useful, because with a "signature" file empty commits are not possible at all.